requirements.txt
requests
pillow
pynput
customtkinter